Struct xmz_mod_touch_server::kombisensor::Sensor

pub struct Sensor { /* fields omitted */ }

Representation der Firmware Daten einer Messzelle

Firmware Version: 0.14.0

Methods

impl Sensor

fn new() -> Self

Erzeugt eine neue Sensor Instanz

Examples

use xmz_mod_touch_server::kombisensor::Sensor;

let sensor = Sensor::new();

fn new_with_type(sensor_type: SensorType) -> Self

Erzeugt eine neue Sensor Instanz vom gegebenen Typ

Parameters

sensor_type - Sensor Typ

Examples

use xmz_mod_touch_server::kombisensor::{Sensor, SensorType};

let sensor_sim_no2_fix = Sensor::new_with_type(SensorType::SimulationNO2Fix);
assert_eq!(sensor_sim_no2_fix.get_concentration(), 20.0)

fn get_adc_value(&self) -> u16

Liefert den aktuell ausgelsenen ADC Wert zurück

Ist der Wert 0 dann deutete das darauf hin das der Sensor noch nicht über Modbus ausgelesen wurde.

Examples

use xmz_mod_touch_server::kombisensor::{Sensor, SensorType};
let sensor = Sensor::new();

assert_eq!(sensor.get_adc_value(), 0);

fn get_min_value(&self) -> u16

Liefert den minimal Wert der für den Sensor konfiguriert wurde

Examples

use xmz_mod_touch_server::kombisensor::Sensor;
let sensor = Sensor::new();

assert_eq!(sensor.get_min_value(), 0);

fn get_max_value(&self) -> u16

Liefert den maximal Wert der für den Sensor konfiguriert wurde

Ist der Wert 0 dann deutete das darauf hin das der Sensor noch nicht über Modbus ausgelesen wurde.

Examples

use xmz_mod_touch_server::kombisensor::Sensor;
let sensor = Sensor::new();

assert_eq!(sensor.get_max_value(), 0);

fn get_adc_value_at_nullgas(&self) -> u16

Liefert den ADC Wert der bei der Kalibration, mit Nullgas, im Sensor gespeichert wurde

Ist der Wert 0 dann deutete das darauf hin das der Sensor noch nicht über Modbus ausgelesen wurde.

Examples

use xmz_mod_touch_server::kombisensor::Sensor;
let sensor = Sensor::new();

assert_eq!(sensor.get_adc_value_at_nullgas(), 0);

fn get_adc_value_at_messgas(&self) -> u16

Liefert den ADC Wert der bei der Kalibration, mit Messgas, im Sensor gespeichert wurde

Ist der Wert 0 dann deutete das darauf hin das der Sensor noch nicht über Modbus ausgelesen wurde.

Examples

use xmz_mod_touch_server::kombisensor::Sensor;
let sensor = Sensor::new();

assert_eq!(sensor.get_adc_value_at_messgas(), 0);

fn get_concentration_at_nullgas(&self) -> u16

Liefert die Konzentration des Gases, die bei der Kalibration, mit Nullgas, im Sensor gespeichert wurde

Examples

use xmz_mod_touch_server::kombisensor::Sensor;
let sensor = Sensor::new();

assert_eq!(sensor.get_concentration_at_nullgas(), 0);

fn get_concentration_at_messgas(&self) -> u16

Liefert die Konzentration des Gases, die bei der Kalibration, mit Messgas, im Sensor gespeichert wurde

Ist der Wert 0 dann deutete das darauf hin das der Sensor noch nicht über Modbus ausgelesen wurde.

Examples

use xmz_mod_touch_server::kombisensor::Sensor;
let sensor = Sensor::new();

assert_eq!(sensor.get_concentration_at_messgas(), 0);

fn get_config(&self) -> u16

Liefert Config Register des Sensors

Examples

use xmz_mod_touch_server::kombisensor::{Sensor, SensorType};
let sensor = Sensor::new_with_type(SensorType::SimulationNO2);

assert_eq!(sensor.get_config(), 0);

fn get_sensor_type(&self) -> SensorType

Liefert den Typen des Sensors

Examples

use xmz_mod_touch_server::kombisensor::{Sensor, SensorType};
let sensor = Sensor::new();

assert_eq!(sensor.get_sensor_type(), SensorType::NemotoNO2);

fn get_si(&self) -> SI

Liefert die SI Einheit des Sensors

Examples

use xmz_mod_touch_server::kombisensor::{Sensor, SI};
let sensor = Sensor::new();

assert_eq!(sensor.get_si(), SI::ppm);

fn get_error_count(&self) -> u64

Liefert den Stand des Fehlerzählers

Examples

use xmz_mod_touch_server::kombisensor::Sensor;
let sensor = Sensor::new();

assert_eq!(sensor.get_error_count(), 0);

fn inc_error_count(&mut self)

Inkrementiert den Fehlerzähler des Sensors

Examples

use xmz_mod_touch_server::kombisensor::Sensor;
let mut sensor = Sensor::new();

assert_eq!(sensor.get_error_count(), 0);
sensor.inc_error_count();
assert_eq!(sensor.get_error_count(), 1);

fn reset_error_count(&mut self)

Setzt den Fehlerzähler des Sensors wieder auf Nullgas

# Examples

use xmz_mod_touch_server::kombisensor::Sensor;

let mut sensor = Sensor::new();
assert_eq!(sensor.get_error_count(), 0);
sensor.inc_error_count();
assert_eq!(sensor.get_error_count(), 1);

sensor.reset_error_count();
assert_eq!(sensor.get_error_count(), 0);

fn direct_value_reached(&self) -> bool

Direktwert überschritten?

Liefert ein boolen true wenn der konfigurierte Direktwert überschritten wurden

# Examples

use xmz_mod_touch_server::kombisensor::{Sensor, SensorType};
let sensor = Sensor::new_with_type(SensorType::SimulationNO2Fix);

assert_eq!(sensor.direct_value_reached(), true)

fn get_concentration(&self) -> f64

Direktwert

Examples

use xmz_mod_touch_server::kombisensor::{Sensor, SensorType};

let sensor = Sensor::new_with_type(SensorType::NemotoNO2);
assert_eq!(sensor.get_concentration(), 0.0);

fn get_concentration_average_15min(&self) -> f64

Mittelwert 15 Minuten

Examples

use xmz_mod_touch_server::kombisensor::{Sensor, SensorType};

let sensor = Sensor::new_with_type(SensorType::NemotoNO2);
assert_eq!(sensor.get_concentration_average_15min(), 0.0);

fn get_mv(&self) -> u16

Liefert den berechneten milli Volt Wert

# Examples

use xmz_mod_touch_server::kombisensor::{Sensor, SensorType};

let sensor = Sensor::new_with_type(SensorType::NemotoNO2);
assert_eq!(sensor.get_mv(), 0);

fn is_enabled(&self) -> bool

Liefert true oder false je nachdem ob der Sensor aktiv Ist

Examples

use xmz_mod_touch_server::kombisensor::{Sensor, SensorType};

let sensor = Sensor::new_with_type(SensorType::SimulationNO2);
assert_eq!(sensor.is_enabled(), false);

fn set_adc_value(&mut self, adc_value: u16)

Setzt den ADC Wert manuell von Hand

Diese Funktion sollte nur zu Test oder Simulationszwecken nötig sein

Examples

use xmz_mod_touch_server::kombisensor::{Sensor, SensorType};
let mut sensor_sim_no2_fix = Sensor::new_with_type(SensorType::SimulationNO2Fix);

sensor_sim_no2_fix.set_adc_value(386);
assert_eq!(sensor_sim_no2_fix.get_concentration(), 30.0);

fn set_min_value(&mut self, min_value: u16)

Setzt den minimal Wert der für den Sensor konfiguriert wurde

Parameters

min_value - Minimal Wert

Examples

use xmz_mod_touch_server::kombisensor::Sensor;
let mut sensor = Sensor::new();

sensor.set_min_value(1);
assert_eq!(sensor.get_min_value(), 1);

fn set_max_value(&mut self, max_value: u16)

Setzt den maximal Wert der für den Sensor konfiguriert wurde

Parameters

max_value - Maximal Wert

Examples

use xmz_mod_touch_server::kombisensor::Sensor;
let mut sensor = Sensor::new();

sensor.set_max_value(100);
assert_eq!(sensor.get_max_value(), 100);

fn set_adc_value_at_nullgas(&mut self, adc_value_at_nullgas: u16)

Setzt den ADC Wert der bei der Kalibration, mit Nullgas, im Sensor gespeichert wurde

Parameters

adc_value_at_nullgas - ADC Wert bie der Nullgas Kalibration

Examples

use xmz_mod_touch_server::kombisensor::Sensor;
let mut sensor = Sensor::new();

sensor.set_adc_value_at_nullgas(100);
assert_eq!(sensor.get_adc_value_at_nullgas(), 100);

fn set_adc_value_at_messgas(&mut self, adc_value_at_messgas: u16)

Setzt den ADC Wert der bei der Kalibration, mit Messgas, im Sensor gespeichert wurde

Parameters

adc_value_at_messgas - ADC Wert bie der Messgas Kalibration

Examples

use xmz_mod_touch_server::kombisensor::Sensor;
let mut sensor = Sensor::new();

sensor.set_adc_value_at_messgas(100);
assert_eq!(sensor.get_adc_value_at_messgas(), 100);

fn set_concentration_at_nullgas(&mut self, concentration_at_nullgas: u16)

Setzt die Konzentration des Gases, die bei der Kalibration, mit Nullgas, im Sensor gespeichert wurde

Parameters

concentration_at_nullgas - Konzentration bei Nullgas

Examples

use xmz_mod_touch_server::kombisensor::Sensor;
let mut sensor = Sensor::new();

sensor.set_concentration_at_nullgas(100);
assert_eq!(sensor.get_concentration_at_nullgas(), 100);

fn set_concentration_at_messgas(&mut self, concentration_at_messgas: u16)

Setzt die Konzentration des Gases, die bei der Kalibration, mit Messgas, im Sensor gespeichert wurde

Parameters

concentration_at_messgas - Konzentration bei Messgas

Examples

use xmz_mod_touch_server::kombisensor::Sensor;
let mut sensor = Sensor::new();

sensor.set_concentration_at_messgas(100);
assert_eq!(sensor.get_concentration_at_messgas(), 100);

fn set_config(&mut self, config: u16)

Setzt Config Register des Sensors

Examples

use xmz_mod_touch_server::kombisensor::{Sensor, SensorType};
let mut sensor = Sensor::new_with_type(SensorType::SimulationNO2);

sensor.set_config(1);
assert_eq!(sensor.get_config(), 1);

fn update_adc_values_average(&mut self)