Struct xmz_mod_touch_server::Kombisensor

pub struct Kombisensor { /* fields omitted */ }

Ein Kombisensor kann n Sensormesszellen enthalten, nomal sind 2 Messzellen (NO2 und CO)

Methods

impl Kombisensor

fn new() -> Self

Erzeugt eine neue Kombisensor Instanz

Return values

Diese Funktion liefert eine neue Kombisensor Instanz

Examples

use xmz_mod_touch_server::Kombisensor;

let kombisensor = Kombisensor::new();
assert_eq!(kombisensor.get_sensors().len(), 2); // 2 Kombisensoren sind default

fn new_with_type(kombisensor_type: KombisensorType) -> Self

Erzeugt eine spezielle Kombisensor Instanz

Parameters

kombisensor_type - Typ des Kombisensors

Return values

Diese Funktion liefert eine neue Kombisensor Instanz vom gegebenen Typ

Examples

use xmz_mod_touch_server::{Kombisensor, KombisensorType};
let kombisensor = Kombisensor::new_with_type(KombisensorType::RAGas);

assert_eq!(kombisensor.get_modbus_device(), "/dev/ttyS0".to_string());

fn get_firmware_version(&self) -> String

Get firmware version

Return values

Liefert die aktuelle Firmware Version die aus dem Sensor ausgelesen wurde, eine Version "0.0.0" deutet darauf hin das der Kombisensor noch nie über Modbus ausgelesen wurde.

Examples

use xmz_mod_touch_server::Kombisensor;

let kombisensor = Kombisensor::new();
assert_eq!(kombisensor.get_firmware_version(), "0.0.0");

fn set_firmware_version(&mut self, firmware_version: String)

Set firmware version

Parameters

firmware_version - String mit der neuen Firmware Version

Examples

use xmz_mod_touch_server::Kombisensor;

let mut kombisensor = Kombisensor::new();

assert_eq!(kombisensor.get_firmware_version(), "0.0.0");
kombisensor.set_firmware_version("9.9.99".to_string());
assert_eq!(kombisensor.get_firmware_version(), "9.9.99");

fn get_modbus_address(&self) -> u8

Get modbus_address

Return values

Liefert die aktuelle Firmware Version die aus dem Sensor ausgelesen wurde, eine Version "0.0.0" deutet darauf hin das der Kombisensor noch nie über Modbus ausgelesen wurde.

Examples

use xmz_mod_touch_server::Kombisensor;

let kombisensor = Kombisensor::new();
assert_eq!(kombisensor.get_modbus_address(), 247);

fn set_modbus_address(&mut self, modbus_address: u8)

Set modbus_address

Parameters

modbus_address - u8 mit der neuen Modbus Adresse

Examples

use xmz_mod_touch_server::Kombisensor;

let mut kombisensor = Kombisensor::new();
assert_eq!(kombisensor.get_modbus_address(), 247);
kombisensor.set_modbus_address(1);
assert_eq!(kombisensor.get_modbus_address(), 1);

fn get_modbus_device(&self) -> String

Get modbus_device

Return values

Liefert die Modbus Device Adresse als String

Examples

use xmz_mod_touch_server::Kombisensor;

let kombisensor = Kombisensor::new();
assert_eq!(kombisensor.get_modbus_device(), "/dev/ttyUSB0".to_string());

fn set_modbus_device(&mut self, modbus_device: String)

Set modbus_device

Parameters

modbus_device - String mit der neuen Modbus Device Adresse

Examples

use xmz_mod_touch_server::Kombisensor;
let mut kombisensor = Kombisensor::new();

kombisensor.set_modbus_device("/dev/ttyS0".to_string());
assert_eq!(kombisensor.get_modbus_device(), "/dev/ttyS0".to_string());

fn get_sensors(&self) -> &Vec<Sensor>

Liefert eine Referenz auf einen Vector mit den Sensoren

Examples

use xmz_mod_touch_server::Kombisensor;

let kombisensor = Kombisensor::new();
assert_eq!(kombisensor.get_sensors().len(), 2); // 2 Kombisensoren sind default

fn get_sensors_mut(&mut self) -> &mut Vec<Sensor>

Liefert eine mutable Referenz auf einen Vector mit den Sensoren

Examples

use xmz_mod_touch_server::Kombisensor;

let mut kombisensor = Kombisensor::new();
assert_eq!(kombisensor.get_sensors_mut().len(), 2); // 2 Kombisensoren sind default

fn get_sensor(&self, id: usize) -> Option<&Sensor>

Sensor Messzellen des Kombisensors

Return values

Liefert ein Option Type, mit den Sensoren

Parameters

id - Id der Sensormesszelle

Examples

use xmz_mod_touch_server::Kombisensor;

let kombisensor = Kombisensor::new();
assert!(kombisensor.get_sensor(0).is_some());

fn get_sensor_mut(&mut self, id: usize) -> Option<&mut Sensor>

Mutable Refernz auf Sensor Messzellen des Kombisensors

Return values

Liefert ein Option Type, mit den mut Sensoren

Parameters

id - Id der Sensormesszelle

Examples

use xmz_mod_touch_server::Kombisensor;

let mut kombisensor = Kombisensor::new();
assert!(kombisensor.get_sensor_mut(0).is_some());

fn get_error_count(&self) -> u64

Get Error Counter

Return values

Liefert den aktuellen Stand des Fehlerzählers

Examples

use xmz_mod_touch_server::Kombisensor;

let kombisensor = Kombisensor::new();
assert_eq!(kombisensor.get_error_count(), 0);

fn inc_error_count(&mut self)

Erhöht den Error Counter

Examples

use xmz_mod_touch_server::Kombisensor;

let mut kombisensor = Kombisensor::new();

assert_eq!(kombisensor.get_error_count(), 0);
kombisensor.inc_error_count();
assert_eq!(kombisensor.get_error_count(), 1);

fn reset_error_count(&mut self)

Reset den Error Counter

Examples

use xmz_mod_touch_server::Kombisensor;
let mut kombisensor = Kombisensor::new();
kombisensor.inc_error_count();
assert_eq!(kombisensor.get_error_count(), 1);

kombisensor.reset_error_count();
assert_eq!(kombisensor.get_error_count(), 0);

fn get_from_modbus(&mut self) -> Result<()>

Fragt die Daten des Kombisensors via Modbus ab

Return values

Die Funktion liefert ein Result

Examples

use xmz_mod_touch_server::Kombisensor;
let mut kombisensor = Kombisensor::new();

assert!(kombisensor.get_from_modbus().is_ok());