Struct xmz_mod_touch_server::xmz_mod_touch_server::XMZModTouchServer

pub struct XMZModTouchServer {
    pub exceptions: HashSet<Exception>,
    // some fields omitted
}

Der XMZModTouchServer kann n Zonen enthalten

Fields
exceptions: HashSet<Exception>
Methods

impl XMZModTouchServer

fn new() -> Self

Erzeugt eine neue XMZModTouchServer Instanz

Return values

Diese Funktion liefert eine neue XMZModTouchServer Instanz

Parameters

Examples

use xmz_mod_touch_server::XMZModTouchServer;

let xmz_mod_touch_server = XMZModTouchServer::new();
assert_eq!(xmz_mod_touch_server.get_version(), env!("CARGO_PKG_VERSION").to_string());

fn new_from_config() -> Result<XMZModTouchServer>

Serverinstanz aus Konfigurationsdatei erstellen

Return values

Diese Funktion liefert ein Result. Das Result enthält die Server Instanz, oder ein Error, wenn die Konfiguration nicht ausgelesen werden konnte.

Examples

use xmz_mod_touch_server::XMZModTouchServer;

let xmz_mod_touch_server = XMZModTouchServer::new_from_config();

fn check(&mut self)

Check Funktion des XMZModTouchServer

Hier werden die Zonen durchlaufen, und deren check() Funktion aufgerufen.

Examples

use xmz_mod_touch_server::XMZModTouchServer;

let mut xmz_mod_touch_server = XMZModTouchServer::new();
xmz_mod_touch_server.check();

fn update(&mut self)

Update Funktion des XMZModTouchServer

Hier werden die Zonen durchlaufen, und deren update() Funktion aufgerufen.

Examples

use xmz_mod_touch_server::XMZModTouchServer;

let mut xmz_mod_touch_server = XMZModTouchServer::new();
xmz_mod_touch_server.update();

fn basic_configuration(&mut self)

basic_configuration - Grundkonfiguration/ Grundeistellungen der LEDs und Relais

Examples

use xmz_mod_touch_server::XMZModTouchServer;

let mut xmz_mod_touch_server = XMZModTouchServer::new();
xmz_mod_touch_server.basic_configuration();

fn get_version(&self) -> String

Liefert die Versionsnummer des XMZModTouchServer's

Die Versionsnummer entspricht der Crate Versionsnummer, wird aus dieser automatisch gebildet.

Return values

Diese Funktion liefert eine neue XMZModTouchServer Instanz

Parameters

Examples

use xmz_mod_touch_server::XMZModTouchServer;

let xmz_mod_touch_server = XMZModTouchServer::new();
assert_eq!(xmz_mod_touch_server.get_version(), env!("CARGO_PKG_VERSION").to_string());

fn get_exceptions(&self) -> &HashSet<Exception>

Liefert eine Refernz auf die Exception des Servers

Examples

use xmz_mod_touch_server::XMZModTouchServer;

let mut xmz_mod_touch_server = XMZModTouchServer::new();
assert_eq!(xmz_mod_touch_server.get_exceptions().len(), 0);

fn get_exception(&self, id: usize) -> Option<&Exception>

Finde eine Exception

Return values

Liefert ein Option Type der entweder eine Refernz auf die Exception des Servers oder None enthält, wenn eine Exception mit dieser Id nicht existiert

Examples

use xmz_mod_touch_server::XMZModTouchServer;
let mut xmz_mod_touch_server = XMZModTouchServer::new();
xmz_mod_touch_server.add_exception(Exception::new(ExceptionType::WartungsintervalReached));

assert!(xmz_mod_touch_server.get_exception(0).is_some());

fn add_exception(&mut self, exception: Exception)

Fügt eine Exception hinzu

Parameters

exception - Exception die hinzugefügt werden soll

Examples

use xmz_mod_touch_server::{XMZModTouchServer, Exception, ExceptionType};

let mut xmz_mod_touch_server = XMZModTouchServer::new();
assert_eq!(xmz_mod_touch_server.get_exceptions().len(), 0);

xmz_mod_touch_server.add_exception(Exception::new(ExceptionType::WartungsintervalReached));
assert_eq!(xmz_mod_touch_server.get_exceptions().len(), 1);

// if the exception is alreddy present, dont insert again
xmz_mod_touch_server.add_exception(Exception::new(ExceptionType::WartungsintervalReached));
assert_eq!(xmz_mod_touch_server.get_exceptions().len(), 1);

fn get_zones(&self) -> &Vec<Zone>

Zonen des Servers

Return values

Liefert eine Refernz auf die Zonen des Servers

Examples

use xmz_mod_touch_server::XMZModTouchServer;

let xmz_mod_touch_server = XMZModTouchServer::new();
assert_eq!(xmz_mod_touch_server.get_zones().len(), 0); // Eine Zone default

fn get_zones_mut(&mut self) -> &mut Vec<Zone>

Mutable Refernz auf die Zonen des Servers

Return values

Liefert eine mutable Refernz auf die Zonen des Servers

Examples

use xmz_mod_touch_server::XMZModTouchServer;

let mut xmz_mod_touch_server = XMZModTouchServer::new();
assert_eq!(xmz_mod_touch_server.get_zones_mut().len(), 0); // Eine Zone default

fn get_zone(&self, id: usize) -> Option<&Zone>

Finde Zone

Return values

Liefert ein Option Typen, der eine Refernz auf die gesucht Zone oder None enthält

Parameters

id - Id der gesuchten Zone

Examples

use xmz_mod_touch_server::XMZModTouchServer;

let xmz_mod_touch_server = XMZModTouchServer::new();
assert!(xmz_mod_touch_server.get_zone(0).is_none());

fn get_zone_mut(&mut self, id: usize) -> Option<&mut Zone>

Finde mut Referenz auf Zone

Return values

Liefert ein Option Typen, der eine mutable Refernz auf die gesucht Zone oder None enthält

Parameters

id - Id der gesuchten Zone

Examples

use xmz_mod_touch_server::XMZModTouchServer;

let xmz_mod_touch_server = XMZModTouchServer::new();
assert!(xmz_mod_touch_server.get_zone_mut(0).is_none());

fn add_zone(&mut self)

Erzeugt eine neu Zone

Examples

use xmz_mod_touch_server::XMZModTouchServer;
let mut xmz_mod_touch_server = XMZModTouchServer::new();
assert!(xmz_mod_touch_server.get_zone(0).is_none());

xmz_mod_touch_server.add_zone();
assert!(xmz_mod_touch_server.get_zone(0).is_some());

fn get_leds(&self) -> &ShiftRegister

Referenz auf die LED's ShiftRegister

Examples

use xmz_mod_touch_server::XMZModTouchServer;

let xmz_mod_touch_server = XMZModTouchServer::new();
let server_leds = xmz_mod_touch_server.get_leds();
assert_eq!(server_leds.data, 0);

fn get_leds_mut(&mut self) -> &mut ShiftRegister

Mutable Referenz auf die LED's ShiftRegister

Examples

use xmz_mod_touch_server::XMZModTouchServer;

let mut xmz_mod_touch_server = XMZModTouchServer::new();
let mut server_leds = xmz_mod_touch_server.get_leds();
assert_eq!(server_leds.data, 0);

fn get_relais(&self) -> &ShiftRegister

Referenz auf die Relais ShiftRegister

Examples

use xmz_mod_touch_server::XMZModTouchServer;

let xmz_mod_touch_server = XMZModTouchServer::new();
let server_relais = xmz_mod_touch_server.get_relais();
assert_eq!(server_relais.data, 0);

fn get_relais_mut(&mut self) -> &mut ShiftRegister

Mutable Referenz auf die Relais ShiftRegister

Examples

use xmz_mod_touch_server::XMZModTouchServer;

let mut xmz_mod_touch_server = XMZModTouchServer::new();
let mut server_relais = xmz_mod_touch_server.get_relais();
assert_eq!(server_relais.data, 0);

fn uptime(&self) -> Duration

Uptime des Servers

Return values

uptime - Die Uptime des Servers

Examples

use xmz_mod_touch_server::XMZModTouchServer;

let xmz_mod_touch_server = XMZModTouchServer::new();
::std::thread::sleep(::std::time::Duration::from_millis(10));
assert!(xmz_mod_touch_server.uptime().num_milliseconds() >= 10);