Struct xmz_mod_touch_server::xmz_mod_touch_server::Zone

pub struct Zone { /* fields omitted */ }

Eine Zone kann n Kombisensoren enthalten

Methods

impl Zone

fn new() -> Self

Erstellt eine neue Zone

Return values

Eine neue Zone

Examples

use xmz_mod_touch_server::Zone;

let zone = Zone::new();

fn get_kombisensors(&self) -> &Vec<Kombisensor>

Liefert eine Refernz auf einen Vector der die Kombisensoren der Zone entält

Examples

use xmz_mod_touch_server::Zone;

let zone = Zone::new();
assert_eq!(zone.get_kombisensors().len(), 2); // 2 Kombisensoren sind default

fn get_kombisensors_mut(&mut self) -> &mut Vec<Kombisensor>

Liefert eine mutable Refernz auf einen Vector der die Kombisensoren der Zone entält

Examples

use xmz_mod_touch_server::Zone;

let mut zone = Zone::new();
assert_eq!(zone.get_kombisensors_mut().len(), 2); // 2 Kombisensoren sind default

fn get_kombisensor(&self, id: usize) -> Option<&Kombisensor>

Finde Kombisensor by Id

Return values

Liefert eine Referenz auf einen Kombisensor, wenn er vorhanden ist oder None, wenn der Kombisensor nicht Teil der Zone ist.

Parameters

id - Id im Vector Kombisensoren (kombisensors) der Zone

Examples

use xmz_mod_touch_server::Zone;

let zone = Zone::new();
assert!(zone.get_kombisensor(0).is_some());

fn get_kombisensor_mut(&mut self, id: usize) -> Option<&mut Kombisensor>

Finde mutablen Kombisensor by Id

Return values

Liefert eine mutable Referenz auf einen Kombisensor, wenn er vorhanden ist oder None, wenn der Kombisensor nicht Teil der Zone ist.

Parameters

id - Id im Vector Kombisensoren (kombisensors) der Zone

Examples

use xmz_mod_touch_server::Zone;

let mut zone = Zone::new();
assert!(zone.get_kombisensor_mut(0).is_some());